Create downloads_organizer_v2.py
Decided to keep the original as the most lightweight as it can be, because sometimes it's just need to do its job simple as that.
HOWEVER the v2 does logging, checks for unique filenames, has CLI arguments and config file  support. For now. Also it supports Linux, Mac and Windows.